signals: move ParseSignal to containerd package
