+++ signal_map_unix.go
@@ -16,7 +16,7 @@
    limitations under the License.
 */
 
-package commands
+package containerd
 
 import (
 	"syscall"
